+++ include/t1m/internal/tensor.h
@@ -1,4 +1,5 @@
 #pragma once
+
 #include <array>
 #include <cstddef>
